fix tests and add logger
